Create QnABot_DaVinci
This is a basic da vinci based qna bot but as you may have guessed you have limited amount of runs for the bot and without prompt engineering it is very useless. Future bots will be variations of these to understad how prompt engineering may help and why openai is a money hungry company.